Save unflagged visibilities instead of flagged visbilities
Model and dirty visibilities have been saved with the flagging applied in the past. I've heard from multiple people that we'd like to have model/dirty vis *without* the flagging applied for ease in rerunning. The vis flag array is still exported, so flagged visibilities can still be made after the fact.